# src/main/resources/swagger/swagger.yml
openapi: 3.0.1
info:
  title: Interview Practice API Docs
  version: "1.0.0"
  description: Custom API documentation using swagger.yml file
paths:
...
  /api/payments/user/{userId}:
    get:
      tags: [ payment-controller ]
      summary: Get Payment History By User
      description: Fetch payment history for a specific user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment history fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentResponse"


  # =======================
  # Refund Controller
  # =======================
...
  /api/wallet/{userId}/credit:
    post:
      tags: [ wallet-controller ]
      summary: Credit Wallet
      description: Add amount to wallet
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
        - in: query
          name: amount
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Wallet credited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletResponse"
  /api/wallet/{userId}/debit:
    post:
      tags: [ wallet-controller ]
      summary: Debit Wallet
      description: Debit amount from wallet
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
        - in: query
          name: amount
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Wallet debited or failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletResponse"
...